feat: since we cap c==16 we may as well use uint16

+++ ecc/bls24-315/multiexp_jacobian.go
@@ -20,7 +20,7 @@ func processChunkG1Jacobian[B ibg1JacExtended](chunk uint64,
 	chRes chan<- g1JacExtended,
 	c uint64,
 	points []G1Affine,
-	digits []uint32) {
+	digits []uint16) {
 
 	var buckets B
 	for i := 0; i < len(buckets); i++ {
@@ -99,7 +99,7 @@ func processChunkG2Jacobian[B ibg2JacExtended](chunk uint64,
 	chRes chan<- g2JacExtended,
 	c uint64,
 	points []G2Affine,
-	digits []uint32) {
+	digits []uint16) {
 
 	var buckets B
 	for i := 0; i < len(buckets); i++ {
